API_UPDATE/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
tensorflow==2.17.0
pillow==9.5.0
numpy==1.23.5
python-multipart==0.0.6
python-jose==3.3.0
python-dotenv==1.0.0
pandas==2.0.1
scikit-learn==1.2.2
opencv-python-headless==4.7.0.72
